rename VectorXZ.isZeroLength(); dot/lengthSquared return doubles; make semantics of toQuaternion consistent with rotate, log attempts to normalize zero, remove trig from clampElliptical, add firstQuadrant/isFirstQuadrant

+++ src/jme3utilities/navigation/NavVertex.java
@@ -184,7 +184,7 @@ public class NavVertex
         float maxDot = -2f * direction.length();
         for (NavArc arc : arcs) {
             VectorXZ horizontalDirection = arc.getHorizontalDirection();
-            float dot = horizontalDirection.dot(direction);
+            float dot = (float) horizontalDirection.dot(direction);
             if (dot > maxDot) {
                 result = arc;
                 maxDot = dot;
